+++ app/src/main/java/com/esaudev/wizeline/data/remote/responses/AskResponse.kt
@@ -1,0 +1,7 @@
+package com.esaudev.wizeline.data.remote.responses
+
+data class AskResponse(
+    val amount: String,
+    val book: String,
+    val price: String
+)
