+++ app/src/main/java/com/esaudev/wizeline/data/remote/responses/BidResponse.kt
@@ -1,0 +1,7 @@
+package com.esaudev.wizeline.data.remote.responses
+
+data class BidResponse(
+    val amount: String,
+    val book: String,
+    val price: String
+)
