+++ app/src/main/java/com/esaudev/wizeline/data/remote/responses/TickerPayload.kt
@@ -1,0 +1,13 @@
+package com.esaudev.wizeline.data.remote.responses
+
+data class TickerPayload(
+    val ask: String,
+    val bid: String,
+    val book: String,
+    val created_at: String,
+    val high: String,
+    val last: String,
+    val low: String,
+    val volume: String,
+    val vwap: String
+)
